requirements.txt
fastapi==0.104.0
uvicorn==0.20.0
jinja2==3.0.0
python-multipart==0.0.6
SQLAlchemy==1.4.39
FastAPI-SQLAlchemy==0.2.1
requests==2.32.2
psycopg2==2.9.9
python-dotenv==1.0.1